+++ fabric/src/main/java/com/lx862/jcm/mod/block/MTREnquiryMachineWall.java
@@ -19,7 +19,7 @@ public class MTREnquiryMachineWall extends WallAttachedBlock implements EnquiryM
 
     @Override
     public ActionResult onUse2(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
-        if(!world.isClient()) enquiry(EnquiryScreenType.CLASSIC, world, player);
+        if(!world.isClient()) enquiry(EnquiryScreenType.CLASSIC, pos, world, player);
         return ActionResult.SUCCESS;
     }
 }