+++ fabric/src/main/java/com/lx862/jcm/mod/network/gui/EnquiryUpdateGUIPacket.java
@@ -2,6 +2,7 @@ package com.lx862.jcm.mod.network.gui;
 
 import com.lx862.jcm.mod.data.EnquiryScreenType;
 import com.lx862.jcm.mod.data.TransactionEntry;
+import org.mtr.mapping.holder.BlockPos;
 import org.mtr.mapping.registry.PacketHandler;
 import org.mtr.mapping.tool.PacketBufferReceiver;
 import org.mtr.mapping.tool.PacketBufferSender;
@@ -15,12 +16,14 @@ public class EnquiryUpdateGUIPacket extends PacketHandler {
 
     private final EnquiryScreenType type;
     private final List<TransactionEntry> entries;
+    private final BlockPos pos;
     private final int entryCount;
     private final int remainingBalance;
 
     public EnquiryUpdateGUIPacket(PacketBufferReceiver packetBufferReceiver) {
         this.entries = new ArrayList<>();
         this.type = EnquiryScreenType.valueOf(packetBufferReceiver.readString());
+        this.pos = BlockPos.fromLong(packetBufferReceiver.readLong());
         this.remainingBalance = packetBufferReceiver.readInt();
 
         this.entryCount = packetBufferReceiver.readInt();
@@ -32,16 +35,18 @@ public class EnquiryUpdateGUIPacket extends PacketHandler {
         }
     }
 
-    public EnquiryUpdateGUIPacket(EnquiryScreenType type, List<TransactionEntry> entries, int remainingBalance) {
+    public EnquiryUpdateGUIPacket(EnquiryScreenType type, BlockPos pos, List<TransactionEntry> entries, int remainingBalance) {
         this.type = type;
         this.entries = entries.stream().sorted((a, b) -> (int)(b.time - a.time)).collect(Collectors.toList());
         this.entryCount = entries.size();
+        this.pos = pos;
         this.remainingBalance = remainingBalance;
     }
 
     @Override
     public void write(PacketBufferSender packetBufferSender) {
         packetBufferSender.writeString(type.toString());
+        packetBufferSender.writeLong(pos.asLong());
         packetBufferSender.writeInt(remainingBalance);
         packetBufferSender.writeInt(entries.size());
 
@@ -54,6 +59,6 @@ public class EnquiryUpdateGUIPacket extends PacketHandler {
 
     @Override
     public void runClient() {
-        ClientHelper.openEnquiryScreen(type, entries, remainingBalance);
+        ClientHelper.openEnquiryScreen(type, pos, entries, remainingBalance);
     }
 }