+++ fabric/src/main/java/com/lx862/jcm/mod/render/gui/screen/RVEnquiryScreen.java
@@ -6,6 +6,7 @@ import com.lx862.jcm.mod.render.GuiHelper;
 import com.lx862.jcm.mod.render.gui.screen.base.AnimatedScreen;
 import com.lx862.jcm.mod.util.TextCategory;
 import com.lx862.jcm.mod.util.TextUtil;
+import org.mtr.mapping.holder.BlockPos;
 import org.mtr.mapping.holder.Identifier;
 import org.mtr.mapping.holder.MinecraftClient;
 import org.mtr.mapping.holder.MutableText;
@@ -17,18 +18,21 @@ import java.text.SimpleDateFormat;
 import java.util.List;
 
 public class RVEnquiryScreen extends AnimatedScreen {
-    private static final Identifier putCardScreen = new Identifier(Constants.MOD_ID, "textures/enquiry/card.png");
-    private static final Identifier balance = new Identifier(Constants.MOD_ID, "textures/enquiry/transactions.png");
-    private static final Identifier font  = new Identifier(Constants.MOD_ID, "font1");
-    private static final Identifier octopuscard  = new Identifier(Constants.MOD_ID, "textures/enquiry/octopus_card.png");
+    private static final Identifier cardScreenTexture = Constants.id("textures/enquiry/card.png");
+    private static final Identifier balanceTexture = Constants.id("textures/enquiry/transactions.png");
+    private static final Identifier octopusCardTexture = Constants.id("textures/enquiry/octopus_card.png");
+    private static final Identifier font = new Identifier("mtr", "mtr");
     private final List<TransactionEntry> entries;
-    private boolean showBalance = false;
-    private long remainingBalance;
+    private final BlockPos pos;
+    private final long remainingBalance;
+    private boolean showBalance;
 
-    public RVEnquiryScreen(List<TransactionEntry> entries, int remainingBalance) {
+    public RVEnquiryScreen(BlockPos pos, List<TransactionEntry> entries, int remainingBalance) {
         super(false);
+        this.pos = pos;
         this.entries = entries;
         this.remainingBalance = remainingBalance;
+        this.showBalance = true; // Simplify mechanics
     }
 
     @Override
@@ -52,21 +56,18 @@ public class RVEnquiryScreen extends AnimatedScreen {
         int rectWidth = 150;
         int rectHeight = 70;
 
-        boolean cursorWithinRectangle = mouseX >= rectX && mouseX <= rectX + rectWidth &&
-                mouseY >= rectY && mouseY <= rectY + rectHeight;
-
+        boolean cursorWithinRectangle = mouseX >= rectX && mouseX <= rectX + rectWidth && mouseY >= rectY && mouseY <= rectY + rectHeight;
         if(cursorWithinRectangle) {
             showBalance = true;
         }
 
         if (!showBalance) {
-            GuiHelper.drawTexture(guiDrawing, putCardScreen, (screenWidth - (int) scaledWidth) / 2, (screenHeight - (int) scaledHeight) / 2, (int) scaledWidth, (int) scaledHeight);
-            GuiHelper.drawTexture(guiDrawing, octopuscard, mouseX, mouseY, 140, 86);
+            GuiHelper.drawTexture(guiDrawing, cardScreenTexture, (screenWidth - (int) scaledWidth) / 2, (screenHeight - (int) scaledHeight) / 2, (int) scaledWidth, (int) scaledHeight);
+            GuiHelper.drawTexture(guiDrawing, octopusCardTexture, mouseX, mouseY, 140, 86);
         } else {
-            GuiHelper.drawTexture(guiDrawing, balance, (screenWidth - (int) scaledWidth) / 2, (screenHeight - (int) scaledHeight) / 2, (int) scaledWidth, (int) scaledHeight);
+            GuiHelper.drawTexture(guiDrawing, balanceTexture, (screenWidth - (int) scaledWidth) / 2, (screenHeight - (int) scaledHeight) / 2, (int) scaledWidth, (int) scaledHeight);
 
             int renderY = startY;
-            int count = 0;
 
             graphicsHolder.push();
             graphicsHolder.scale((float)(double)getWidthMapped() / baseWidth, (float)(double)getWidthMapped() / baseWidth, (float)(double)getWidthMapped() / baseWidth);
@@ -74,11 +75,9 @@ public class RVEnquiryScreen extends AnimatedScreen {
             graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.balance"), font), startX, 20, 0, false, GraphicsHolder.getDefaultLight());
             graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal("Octopus"), font), startX + 305, 75, 0, false, GraphicsHolder.getDefaultLight());
             graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.processor"), font), startX + 305, 85, 0, false, GraphicsHolder.getDefaultLight());
-            graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal("849009"), font), startX + 305, 95, 0, false, GraphicsHolder.getDefaultLight());
 
-            graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.last.a"), font), startX + 305, 115, 0, false, GraphicsHolder.getDefaultLight());
-            graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.last.b"), font), startX + 305, 125, 0, false, GraphicsHolder.getDefaultLight());
-            graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.last.c"), font), startX + 305, 135, 0, false, GraphicsHolder.getDefaultLight());
+            String processorId = String.format("%06d", pos.asLong() % 1000000);
+            graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal(processorId), font), startX + 305, 95, 0, false, GraphicsHolder.getDefaultLight());
 
             for (int i = 0; i < 10; i++) {
                 MutableText renderText = getEntryText(i);
@@ -86,15 +85,13 @@ public class RVEnquiryScreen extends AnimatedScreen {
 
                 graphicsHolder.drawText(TextUtil.withFont(renderText, font), startX, renderY, 0, false, GraphicsHolder.getDefaultLight());
                 renderY += 10;
-                count++;
             }
 
             if (!entries.isEmpty()) {
-                TransactionEntry lastTransactionEntry = entries.get(entries.size() - 1);
                 if (remainingBalance >= 0) {
-                    graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal("$" + String.valueOf(remainingBalance)), font), startX + 270, 20, 0x000000, false, GraphicsHolder.getDefaultLight());
-                } else if (remainingBalance < 0) {
-                    graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal("-$" + String.valueOf(Math.abs(remainingBalance))), font), startX + 270, 20, 0x000000, false, GraphicsHolder.getDefaultLight());
+                    graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal("$" + remainingBalance), font), startX + 270, 20, 0x000000, false, GraphicsHolder.getDefaultLight());
+                } else {
+                    graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal("-$" + Math.abs(remainingBalance)), font), startX + 270, 20, 0x000000, false, GraphicsHolder.getDefaultLight());
                 }
             }
 
@@ -109,6 +106,10 @@ public class RVEnquiryScreen extends AnimatedScreen {
             if (lastDate != 0) {
                 SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yyyy HH:mm");
                 String formattedDate = formatter.format(lastDate);
+
+                graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.last.a"), font), startX + 305, 115, 0, false, GraphicsHolder.getDefaultLight());
+                graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.last.b"), font), startX + 305, 125, 0, false, GraphicsHolder.getDefaultLight());
+                graphicsHolder.drawText(TextUtil.withFont(TextUtil.translatable(TextCategory.GUI, "rvenquiry_screen.last.c"), font), startX + 305, 135, 0, false, GraphicsHolder.getDefaultLight());
                 graphicsHolder.drawText(TextUtil.withFont(TextUtil.literal(formattedDate), font), startX + 305, 145, 0x000000, false, GraphicsHolder.getDefaultLight());
             }
 
@@ -122,8 +123,10 @@ public class RVEnquiryScreen extends AnimatedScreen {
         String renderTextString;
         if (transactionEntry.amount < 0) {
             renderTextString = String.format("%s     %s     -$%.2f", transactionEntry.getFormattedDate(), IGui.formatStationName(transactionEntry.source), Math.abs((double) transactionEntry.amount));
-        } else {
+        } else if (transactionEntry.amount > 0) {
             renderTextString = String.format("%s     %s     +$%.2f", transactionEntry.getFormattedDate(), IGui.formatStationName(transactionEntry.source), (double) transactionEntry.amount);
+        } else {
+            renderTextString = String.format("%s     %s     $%.2f", transactionEntry.getFormattedDate(), IGui.formatStationName(transactionEntry.source), (double) transactionEntry.amount);
         }
         MutableText renderText = TextUtil.literal(renderTextString);
         return renderText;